Extract classes / interfaces
